Moved more utilities around
